+++ package.json
@@ -15,7 +15,8 @@
     "@mui/material": "^5.15.3",
     "next": "14.0.4",
     "react": "^18",
-    "react-dom": "^18"
+    "react-dom": "^18",
+    "sass": "^1.69.7"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.2.0",
