+++ src/App.test.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { render } from '@testing-library/react';
+import App from './App';
+
+test('renders the app', () => {
+  const { container } = render(<App />);
+  expect(container).toBeInTheDocument();
+});
