+++ src/pages/index.js
@@ -1,0 +1,26 @@
+import Head from 'next/head'
+import Image from 'next/image'
+import App from '../App';
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <main>
+      <App />
+        <Image
+                src="/vercel.svg"
+                alt="Vercel Logo"
+                width={100}
+                height={24}
+                priority
+              />
+      </main>
+    </>
+  )
+}
