+++ src/scss/partials/_calculator.scss
